Refactor ssh client getters into root mod
